fix: [5주차 피드백] user_id가 null 값이여서 주로 발생하는 엔티티 관계설정 문제로 테스트코드 에러 나는 부분 수정
- 잔액 충전 : 낙관적 락인데 비관적락 이라고 기술한 부분 수정
