+++ .github/pull_request_template.md
@@ -1,22 +1,34 @@
 ## :pushpin: PR 제목 규칙
-[STEP0X] 이름 - 선택 시나리오 (e-commerce / concert)
+[STEP0X] 이름 - 선택 시나리오 (e-commerce/concert)
 
 ---
+### **핵심 체크리스트** :white_check_mark:
 
-## :clipboard: 핵심 체크리스트 :white_check_mark:
+#### :one: 분산락 적용 (3개)
+- [ ] 적절한 곳에 분산락이 사용되었는가?
+- [ ] 트랜젝션 순서와 락순서가 보장되었는가?
 
-### STEP09 - Concurrency (2개)
-- [ ] 애플리케이션 내에서 발생 가능한 **동시성 문제를 식별**했는가?
-- [ ] 보고서에 DB를 활용한 **동시성 문제 해결 방안**이 포함되어 있는가?
+#### :two: 통합 테스트 (2개)
+- [ ] infrastructure 레이어를 포함하는 통합 테스트가 작성되었는가?
+- [ ] 핵심 기능에 대한 흐름이 테스트에서 검증되었는가?
+- [ ] 동시성을 검증할 수 있는 테스트코드로 작성 되었는가?
+- [ ] Test Container 가 적용 되었는가?
+
+#### :three: Cache 적용 (3개)
+- [ ] 적절하게 Key 적용이 되었는가?
 
 ---
+#### STEP11
+- [ ] Redis 분산락 적용
+- [ ] Test Container 구성
+- [ ] 기능별 통합 테스트
 
-### STEP10 - Finalize (1개)
-- [ ] **동시성 문제를 드러낼 수 있는 통합 테스트**를 작성했는가?
+#### STEP12
+- [ ] 캐시 필요한 부분 분석
+- [ ] redis 기반의 캐시 적용
+- [ ] 성능 개선 등을 포함한 보고서 제출
 
----
-
-## ✍️ 간단 회고 (3줄 이내)
-- **잘한 점**: 
-- **어려웠던 점**: 
-- **다음 시도**: 
+### **간단 회고** (3줄 이내)
+- **잘한 점**:
+- **어려운 점**:
+- **다음 시도**:
