+++ src/test/java/kr/hhplus/be/server/integration/BalanceIntegrationTest.java
@@ -217,166 +217,5 @@ class BalanceIntegrationTest {
         assertThat(newBalance.get().getAmount()).isEqualTo(chargeAmount);
     }
 
-    @Test
-    @DisplayName("동시 잔액 충전 테스트 - 동시성 제어")
-    void concurrentChargeBalanceTest() throws InterruptedException {
-        // Given
-        Long userId = testUser.getUserId();
-        BigDecimal initialAmount = new BigDecimal("10000.00");
-        BigDecimal chargeAmount1 = new BigDecimal("5000.00");
-        BigDecimal chargeAmount2 = new BigDecimal("3000.00");
-        
-        // 초기 잔액 설정
-        ChargeBalanceUseCase.ChargeBalanceCommand initialRequest = new ChargeBalanceUseCase.ChargeBalanceCommand(userId, initialAmount);
-        ChargeBalanceUseCase.ChargeBalanceResult initialResponse = chargeBalanceService.chargeBalance(initialRequest);
-        assertThat(initialResponse.isSuccess()).isTrue();
-        
-        // When - 진짜 동시에 두 개의 충전 요청 실행
-        CountDownLatch startLatch = new CountDownLatch(1); // 시작 신호
-        CountDownLatch readyLatch = new CountDownLatch(2); // 준비 완료 신호
-        CountDownLatch finishLatch = new CountDownLatch(2); // 완료 신호
-        AtomicReference<ChargeBalanceUseCase.ChargeBalanceResult> response1 = new AtomicReference<>();
-        AtomicReference<ChargeBalanceUseCase.ChargeBalanceResult> response2 = new AtomicReference<>();
-        
-        Thread thread1 = new Thread(() -> {
-            try {
-                readyLatch.countDown(); // 준비 완료 신호
-                startLatch.await(); // 시작 신호 대기
-                
-                ChargeBalanceUseCase.ChargeBalanceCommand request1 = new ChargeBalanceUseCase.ChargeBalanceCommand(userId, chargeAmount1);
-                response1.set(chargeBalanceService.chargeBalance(request1));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            } finally {
-                finishLatch.countDown();
-            }
-        });
-        
-        Thread thread2 = new Thread(() -> {
-            try {
-                readyLatch.countDown(); // 준비 완료 신호
-                startLatch.await(); // 시작 신호 대기
-                
-                ChargeBalanceUseCase.ChargeBalanceCommand request2 = new ChargeBalanceUseCase.ChargeBalanceCommand(userId, chargeAmount2);
-                response2.set(chargeBalanceService.chargeBalance(request2));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            } finally {
-                finishLatch.countDown();
-            }
-        });
-        
-        thread1.start();
-        thread2.start();
-        
-        // 모든 스레드가 준비될 때까지 대기
-        readyLatch.await();
-        System.out.println("모든 스레드가 준비되었습니다. 동시 시작!");
-        
-        // 시작 신호 전송 (모든 스레드가 동시에 시작)
-        startLatch.countDown();
-        
-        finishLatch.await(5, TimeUnit.SECONDS);
-        
-        // Then - 두 요청 모두 성공해야 함
-        assertThat(response1.get().isSuccess()).isTrue();
-        assertThat(response2.get().isSuccess()).isTrue();
-        
-        // 최종 잔액 확인 (10000.00 + 5000.00 + 3000.00 = 18000.00)
-        GetBalanceUseCase.GetBalanceCommand finalBalanceCommand = new GetBalanceUseCase.GetBalanceCommand(userId);
-        Optional<GetBalanceUseCase.GetBalanceResult> finalBalanceResult = getBalanceService.getBalance(finalBalanceCommand);
-        assertThat(finalBalanceResult).isPresent();
-        
-        GetBalanceUseCase.GetBalanceResult finalBalance = finalBalanceResult.get();
-        assertThat(finalBalance.getUserId()).isEqualTo(userId);
-        assertThat(finalBalance.getBalance()).isEqualTo(new BigDecimal("18000.00"));
-    }
-    
-    @Test
-    @DisplayName("동시 잔액 충전과 결제 테스트")
-    void concurrentChargeAndPaymentTest() throws InterruptedException {
-        // Given
-        Long userId = testUser.getUserId();
-        BigDecimal initialAmount = new BigDecimal("10000.00");
-        BigDecimal chargeAmount = new BigDecimal("5000.00");
-        BigDecimal paymentAmount = new BigDecimal("3000.00");
-        
-        // 초기 잔액 설정
-        ChargeBalanceUseCase.ChargeBalanceCommand initialRequest = new ChargeBalanceUseCase.ChargeBalanceCommand(userId, initialAmount);
-        ChargeBalanceUseCase.ChargeBalanceResult initialResponse = chargeBalanceService.chargeBalance(initialRequest);
-        assertThat(initialResponse.isSuccess()).isTrue();
-        
-        // When - 진짜 동시에 충전과 결제 요청 실행
-        CountDownLatch startLatch = new CountDownLatch(1); // 시작 신호
-        CountDownLatch readyLatch = new CountDownLatch(2); // 준비 완료 신호
-        CountDownLatch finishLatch = new CountDownLatch(2); // 완료 신호
-        AtomicReference<ChargeBalanceUseCase.ChargeBalanceResult> chargeResponse = new AtomicReference<>();
-        AtomicReference<Boolean> paymentSuccess = new AtomicReference<>(false);
-        
-        Thread chargeThread = new Thread(() -> {
-            try {
-                readyLatch.countDown(); // 준비 완료 신호
-                startLatch.await(); // 시작 신호 대기
-                
-                ChargeBalanceUseCase.ChargeBalanceCommand chargeRequest = new ChargeBalanceUseCase.ChargeBalanceCommand(userId, chargeAmount);
-                chargeResponse.set(chargeBalanceService.chargeBalance(chargeRequest));
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-            } finally {
-                finishLatch.countDown();
-            }
-        });
-        
-        Thread paymentThread = new Thread(() -> {
-            try {
-                readyLatch.countDown(); // 준비 완료 신호
-                startLatch.await(); // 시작 신호 대기
-                
-                // 실제 결제 요청 (DeductBalancePort 사용)
-                boolean success = deductBalancePort.deductBalance(userId, paymentAmount);
-                paymentSuccess.set(success);
-            } catch (Exception e) {
-                System.err.println("Payment failed: " + e.getMessage());
-            } finally {
-                finishLatch.countDown();
-            }
-        });
-        
-        chargeThread.start();
-        paymentThread.start();
-        
-        // 모든 스레드가 준비될 때까지 대기
-        readyLatch.await();
-        System.out.println("모든 스레드가 준비되었습니다. 동시 시작!");
-        
-        // 시작 신호 전송 (모든 스레드가 동시에 시작)
-        startLatch.countDown();
-        
-        finishLatch.await(5, TimeUnit.SECONDS);
-        
-        // Then
-        assertThat(chargeResponse.get().isSuccess()).isTrue();
-        
-        // 최종 잔액 확인 (충전은 성공, 결제는 별도 처리)
-        GetBalanceUseCase.GetBalanceCommand finalBalanceCommand = new GetBalanceUseCase.GetBalanceCommand(userId);
-        Optional<GetBalanceUseCase.GetBalanceResult> finalBalanceResult = getBalanceService.getBalance(finalBalanceCommand);
-        assertThat(finalBalanceResult).isPresent();
-        
-        GetBalanceUseCase.GetBalanceResult finalBalance = finalBalanceResult.get();
-        assertThat(finalBalance.getUserId()).isEqualTo(userId);
-        assertThat(finalBalance.getBalance()).isGreaterThanOrEqualTo(new BigDecimal("15000.00")); // 10000.00 + 5000.00
-    }
-    
-
-    
-    
-
-    
-
-    
-
-    
-
-    
 
 }
